[AOTI][refactor] Split common utils from aoti_runtime/model.h
Summary: Split common utils from aoti_runtime/model.h into a separate header file, because when turning on ABI-compatible mode for JIT Inductor we won't need AOTInductorModel, but we do need some common utils, e.g. RAIIAtenTensorHandle.

[ghstack-poisoned]
