+++ torch/csrc/inductor/aoti_runtime/model.h
@@ -16,23 +16,7 @@
 // C ABI defined in torch/csrc/inductor/aoti_torch/c/shim.h. The same rule
 // applies to other files under torch/csrc/inductor/aoti_runtime/.
 #include <torch/csrc/inductor/aoti_runtime/device_utils.h>
-#include <torch/csrc/inductor/aoti_torch/c/shim.h>
-
-#define AOTI_RUNTIME_CHECK(EXPR, MSG) \
-  do {                                \
-    bool ok = EXPR;                   \
-    if (!ok) {                        \
-      throw std::runtime_error(MSG);  \
-    }                                 \
-  } while (0)
-
-#if defined(__GNUC__) || defined(__clang__)
-#define AOTI_NOINLINE __attribute__((noinline))
-#elif _MSC_VER
-#define AOTI_NOINLINE __declspec(noinline)
-#else
-#define AOTI_NOINLINE
-#endif
+#include <torch/csrc/inductor/aoti_runtime/utils.h>
 
 // At codegen time, we write out a binary file called constants.bin.
 // We then turn the raw binary to an object file that exposes this
@@ -63,93 +47,8 @@ CUDAPtr RAII_cudaMalloc(size_t num_bytes) {
 
 } // anonymous namespace
 
-AOTI_NOINLINE static void throw_exception(
-    const char* call,
-    const char* file,
-    int64_t line) {
-  std::stringstream ss;
-  ss << call << " API call failed at " << file << ", line " << line;
-  throw std::runtime_error(ss.str());
-}
-
-#define AOTI_TORCH_ERROR_CODE_CHECK(call)       \
-  if ((call) != AOTI_TORCH_SUCCESS) {           \
-    throw_exception(#call, __FILE__, __LINE__); \
-  }
-
-using DeleterFnPtr = void (*)(void*);
-
 namespace torch {
 namespace aot_inductor {
-
-inline void noop_deleter(void*) {}
-
-inline void delete_tensor_object(void* ptr) {
-  AOTI_TORCH_ERROR_CODE_CHECK(
-      aoti_torch_delete_tensor_object(reinterpret_cast<AtenTensorHandle>(ptr)));
-}
-
-// RAIIAtenTensorHandle steals the tensor objects created by the libtorch C ABI
-class RAIIAtenTensorHandle {
- public:
-  RAIIAtenTensorHandle() : handle_(nullptr, noop_deleter) {}
-  RAIIAtenTensorHandle(const RAIIAtenTensorHandle& other) = delete;
-  RAIIAtenTensorHandle& operator=(const RAIIAtenTensorHandle& other) = delete;
-
-  // Steal the ownership from another RAIIAtenTensorHandle using std::move
-  RAIIAtenTensorHandle(RAIIAtenTensorHandle&& other) = default;
-  RAIIAtenTensorHandle& operator=(RAIIAtenTensorHandle&& other) = default;
-
-  // Steal the ownership from raw AtenTensorHandle
-  RAIIAtenTensorHandle(AtenTensorHandle handle)
-      : handle_(handle, delete_tensor_object) {}
-
-  ~RAIIAtenTensorHandle() {
-    handle_.reset();
-  }
-
-  // Return a raw AtenTensorHandle to be used by aoti_torch functions
-  // Note: this function does NOT transfer the ownership of the handle
-  operator AtenTensorHandle() const {
-    return handle_.get();
-  }
-
-  AtenTensorHandle release() {
-    return handle_.release();
-  }
-
-  AtenTensorHandle get() const {
-    return handle_.get();
-  }
-
-  void reset() {
-    handle_.reset();
-  }
-
-  int64_t size(int64_t d) {
-    int64_t size;
-    AOTI_TORCH_ERROR_CODE_CHECK(aoti_torch_get_size(handle_.get(), d, &size));
-    return size;
-  }
-
-  int64_t stride(int64_t d) {
-    int64_t stride;
-    AOTI_TORCH_ERROR_CODE_CHECK(
-        aoti_torch_get_stride(handle_.get(), d, &stride));
-    return stride;
-  }
-
-  int64_t storage_offset() {
-    int64_t storage_offset;
-    AOTI_TORCH_ERROR_CODE_CHECK(
-        aoti_torch_get_storage_offset(handle_.get(), &storage_offset));
-    return storage_offset;
-  }
-
- private:
-  std::unique_ptr<AtenTensorOpaque, DeleterFnPtr> handle_;
-};
-
 using ConstantMap = std::unordered_map<std::string, RAIIAtenTensorHandle>;
 
 class ConstantHandle {
@@ -189,19 +88,6 @@ inline const ConstantHandle& unwrap_raii_handle_if_needed(
 // Shouldn't be called.
 inline AtenTensorHandle wrap_with_raii_handle_if_needed(
     const ConstantHandle& handle) = delete;
-
-// Steal the ownership from raw AtenTensorHandle to RAIIAtenTensorHandle
-inline std::vector<RAIIAtenTensorHandle> steal_from_raw_handles_to_raii_handles(
-    AtenTensorHandle* handles,
-    size_t size) {
-  std::vector<RAIIAtenTensorHandle> result;
-  result.reserve(size);
-  for (size_t i = 0; i < size; i++) {
-    result.emplace_back(handles[i]);
-    handles[i] = nullptr;
-  }
-  return result;
-}
 
 // valid device strs are: cpu, cuda, cuda:0, cuda:1, ...
 // Update the list here if more devices are supported in the future
@@ -643,25 +529,6 @@ class AOTInductorModel : public AOTInductorModelBase<AOTInductorModel> {
  private:
   std::unique_ptr<AOTInductorModelKernelsBase> kernels_;
 };
-
-#ifdef USE_CUDA
-class AOTICudaStreamGuard {
- public:
-  AOTICudaStreamGuard(cudaStream_t stream, int32_t device_index) {
-    CUDAStreamGuardHandle ptr;
-    AOTI_TORCH_ERROR_CODE_CHECK(
-        aoti_torch_create_cuda_stream_guard(stream, device_index, &ptr));
-    guard_ =
-        std::unique_ptr<void, std::function<void(void*)>>(ptr, [](void* ptr) {
-          AOTI_TORCH_ERROR_CODE_CHECK(aoti_torch_delete_cuda_stream_guard(
-              reinterpret_cast<CUDAStreamGuardHandle>(ptr)));
-        });
-  }
-
- private:
-  std::unique_ptr<void, std::function<void(void*)>> guard_;
-};
-#endif // USE_CUDA
 
 } // namespace aot_inductor
 } // namespace torch
